Put back react-scripts and renamed scss files to .module.scss
