+++ .env.production
@@ -1,3 +1,3 @@
-ALGOLIA_APPLICATION_ID=29FLVJV5X9
-ALGOLIA_API_KEY=f2534ea79a28d8469f4e81d546297d39
-API_END_POINT_DOMAIN=https://confs-tech-api.herokuapp.com
+REACT_APP_ALGOLIA_APPLICATION_ID=29FLVJV5X9
+REACT_APP_ALGOLIA_API_KEY=f2534ea79a28d8469f4e81d546297d39
+REACT_APP_API_END_POINT_DOMAIN=https://confs-tech-api.herokuapp.com
