+++ package.json
@@ -3,9 +3,10 @@
   "version": "1.0.0",
   "private": true,
   "scripts": {
-    "start": "node scripts/start.js",
-    "build": "node scripts/build.js",
-    "test": "node scripts/test.js --env=jsdom",
+    "start": "react-scripts start",
+    "build": "react-scripts build",
+    "test": "react-scripts test",
+    "eject": "react-scripts eject",
     "lint": "eslint --fix .",
     "reorder-confs": "BABEL_ENV=node babel-node ./scripts/reorder-conferences-by-date.js",
     "precommit": "lint-staged",
@@ -26,28 +27,16 @@
     "algoliasearch": "^4.8.6",
     "autoprefixer": "9.1.4",
     "babel-core": "6.26.3",
-    "babel-eslint": "8.2.6",
-    "babel-jest": "23.4.2",
-    "babel-loader": "7.1.5",
     "babel-preset-react-app": "^3.0.2",
     "babel-runtime": "6.26.0",
-    "case-sensitive-paths-webpack-plugin": "2.1.2",
     "chalk": "2.4.1",
     "classnames": "^2.2.5",
-    "css-loader": "^1.0.0",
     "date-fns": "^2.16.1",
     "dotenv": "6.0.0",
-    "eslint-config-react-app": "^2.0.0",
-    "eslint-plugin-flowtype": "2.50.0",
-    "eslint-plugin-jsx-a11y": "6.4.1",
-    "eslint-plugin-react": "7.11.1",
-    "extract-text-webpack-plugin": "3.0.2",
-    "file-loader": "2.0.0",
     "fs-extra": "7.0.0",
     "gh-pages": "^2.0.0",
-    "html-webpack-plugin": "3.2.0",
-    "jest": "^26.6.3",
     "lodash": "^4.17.4",
+    "node-sass": "^5.0.0",
     "object-assign": "4.1.1",
     "postcss-flexbugs-fixes": "4.1.0",
     "postcss-loader": "3.0.0",
@@ -58,20 +47,12 @@
     "react-dev-utils": "^5.0.2",
     "react-dom": "^17.0.1",
     "react-error-overlay": "^4.0.1",
-    "react-favicon": "0.0.14",
     "react-helmet": "^6.1.0",
     "react-instantsearch-dom": "^6.10.3",
     "react-recaptcha": "^2.3.7",
     "react-router": "^5.2.0",
     "react-router-dom": "^5.2.0",
-    "style-loader": "0.22.1",
-    "sw-precache-webpack-plugin": "0.11.5",
-    "ts-loader": "3.5.0",
-    "tsconfig-paths-webpack-plugin": "^3.2.0",
-    "url-loader": "^1.0.1",
-    "webpack": "^3.10.0",
-    "webpack-dev-server": "^2.9.7",
-    "webpack-manifest-plugin": "^2.0.3",
+    "react-scripts": "^4.0.3",
     "whatwg-fetch": "2.0.4"
   },
   "devDependencies": {
@@ -90,25 +71,18 @@
     "@types/react-recaptcha": "^2.3.3",
     "@types/react-router-dom": "^4.3.0",
     "@types/universal-cookie": "^2.2.0",
-    "@types/webpack": "4.4.11",
     "@typescript-eslint/eslint-plugin": "^4.17.0",
     "@typescript-eslint/parser": "^4.17.0",
     "babel-plugin-transform-es2015-modules-commonjs": "^6.26.0",
     "babel-preset-env": "^1.6.1",
-    "eslint": "^7.21.0",
     "eslint-config-prettier": "^7.1.0",
-    "eslint-loader": "^4.0.2",
     "eslint-plugin-import": "^2.22.1",
     "eslint-plugin-prettier": "^3.3.1",
     "eslint-plugin-react-hooks": "^4.2.0",
-    "fork-ts-checker-webpack-plugin": "^1.4.3",
     "husky": "^4.3.7",
-    "jest-cli": "^23.5.0",
     "jsonlint": "^1.6.3",
     "lint-staged": "^9.2.1",
-    "node-sass": "^4.5.3",
     "prettier": "^2.2.1",
-    "sass-loader": "^7.1.0",
     "source-map-loader": "^0.2.3",
     "tslint": "5.11.0",
     "typescript": "^4.2.3"
@@ -158,6 +132,18 @@
       "webpack",
       "webpack-dev-server",
       "webpack-manifest-plugin"
+    ]
+  },
+  "browserslist": {
+    "production": [
+      ">0.2%",
+      "not dead",
+      "not op_mini all"
+    ],
+    "development": [
+      "last 1 chrome version",
+      "last 1 firefox version",
+      "last 1 safari version"
     ]
   }
 }
