+++ src/StaticPages/NotFoundPage.tsx
@@ -1,4 +1,3 @@
-import React from 'react'
 import { Page } from 'src/components'
 
 export default function NotFoundPage() {
