--- src/components/Banner/Banner.scss
@@ -1,10 +1,0 @@
-@import 'src/components/style/foundation/spacing';
-@import 'src/components/style/foundation/colors';
-
-.Banner {
-  margin: spacing() 0;
-  padding: spacing();
-  background: color(accent-2, pale);
-  border-radius: 3px;
-  position: relative;
-}
