+++ src/components/Card/Card.tsx
@@ -1,6 +1,4 @@
-import React from 'react'
-
-import styles from './Card.scss'
+import styles from './Card.module.scss'
 
 const Card: React.FC = ({ children }) => {
   return <div className={styles.Card}>{children}</div>
