+++ src/components/Divider/Divider.tsx
@@ -1,6 +1,4 @@
-import React from 'react'
-
-import styles from './Divider.scss'
+import styles from './Divider.module.scss'
 
 export default function Divider() {
   return <hr className={styles.Divider} />
