+++ src/components/GithubStar/GithubStar.module.scss
