+++ src/components/Header/index.ts
@@ -1,2 +1,1 @@
 export { default } from './Header'
-export { Props } from './Header'
