+++ src/components/Twitter/Twitter.tsx
@@ -1,5 +1,3 @@
-import React from 'react'
-
 import Link from '../Link'
 
 interface Props {
