+++ src/components/index.ts
@@ -3,7 +3,7 @@ export { default as Divider } from './Divider'
 export { default as Footer } from './Footer'
 export { default as GithubStar } from './GithubStar'
 export { default as Head } from './Head'
-export { default as Header, Props as HeaderProps } from './Header'
+export { default as Header } from './Header'
 export { default as Heading } from './Heading'
 export { default as InputGroup } from './InputGroup'
 export { default as Link } from './Link'
