+++ src/index.tsx
@@ -1,7 +1,6 @@
-import React from 'react'
 import ReactDOM from 'react-dom'
 import { BrowserRouter } from 'react-router-dom'
-import './index.scss'
+import './index.module.scss'
 
 import App from './components/App'
 import { DarkModeContextProvider } from './contexts/DarkModeContext'
