+++ src/scenes/ConferenceList/components/CFPHeader/CFPHeader.module.scss
