+++ src/scenes/ConferenceList/components/CFPHeader/CFPHeader.tsx
@@ -1,7 +1,6 @@
-import React from 'react'
 import { Link, Heading } from 'src/components'
 
-import styles from './CFPHeader.scss'
+import styles from './CFPHeader.module.scss'
 
 const CfpHeader = ({ sortByCfpEndDate, sortBy }: any) => {
   return (
