+++ src/scenes/ConferenceList/components/ConferenceItem/ConferenceItem.module.scss
