+++ src/scenes/ConferenceList/components/ConferenceList/ConferenceList.module.scss
