+++ types/conference.ts
@@ -1,0 +1,18 @@
+export interface Conference {
+  objectID: string
+  name: string
+  online: boolean
+  topics: string[]
+  url: string
+  city: string
+  country: string
+  startDate: string
+  endDate: string
+  twitter: string
+  cfpEndDate: string
+  cfpUrl: string
+  affiliateText: string
+  affiliateUrl: string
+  cocUrl: string
+  offersSignLanguageOrCC: boolean
+}
